Discord Bot/requirements.txt
discord.py==2.1.0
py-cord==2.3.2
firebase==3.0.1
requests==2.28.1
python-firebase==1.2
firebase-admin==6.0.1
pycord==0.1.1